+++ resources/assets/js/app.js
@@ -231,3 +231,6 @@ $(function(){
         }); 
     }); 
 });
+
+// Lity Lightbox
+var lity = require('lity');