+++ resources/lang/de/app.php
@@ -157,4 +157,6 @@ return [
     'project' => 'Projekt',
     'description' => 'Beschreibung',
     'no_data_available_in_the_selected_time_range' => 'Keine Daten im ausgewählten Zeitraum vorhanden.',
+    'overview' => 'Übersicht',
+    'remove' => 'Entfernen',
 ];